Rename package name for annotation.
