+++ src/main/java/net/datafaker/annotations/EmbeddedFake.java
@@ -1,4 +1,4 @@
-package net.datafaker.annotaions;
+package net.datafaker.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
