+++ src/main/java/net/datafaker/annotations/EmbeddedFakeProvider.java
@@ -1,4 +1,4 @@
-package net.datafaker.annotaions;
+package net.datafaker.annotations;
 
 import java.lang.reflect.Field;
