+++ src/main/java/net/datafaker/annotations/FakeResolver.java
@@ -1,4 +1,4 @@
-package net.datafaker.annotaions;
+package net.datafaker.annotations;
 
 import java.util.Locale;
 import java.util.Objects;
