+++ src/main/java/net/datafaker/annotations/ObjectProvider.java
@@ -1,4 +1,4 @@
-package net.datafaker.annotaions;
+package net.datafaker.annotations;
 
 interface ObjectProvider {
