+++ src/test/java/net/datafaker/annotations/FakeRecordProviderTest.java
@@ -1,4 +1,4 @@
-package net.datafaker.annotaions;
+package net.datafaker.annotations;
 
 import java.sql.Timestamp;
